+++ src/database/models.rs
@@ -1,3 +1,5 @@
+mod friend;
 mod user;
 
+pub use friend::FriendModel;
 pub use user::UserModel;